Reorganise github/client and github/release
