+++ pkg/github/client/client_iface.go
@@ -1,4 +1,4 @@
-package githubClient
+package client
 
 import (
 	"github.com/google/go-github/github"
